refact: extrair lógica de deleção para DeleteActions (actor-sheet.mjs)
