+++ packs/efeitos-cardigan/CURRENT
@@ -1,1 +1,1 @@
-MANIFEST-001318
+MANIFEST-001320
