--- packs/efeitos-cardigan/LOG
@@ -1,3 +1,0 @@
-2026/07/02-22:15:18.909 3304 Recovering log #1317
-2026/07/02-22:15:19.475 3304 Delete type=0 #1317
-2026/07/02-22:15:19.475 3304 Delete type=3 #1316
