+++ packs/efeitos-cardigan/LOG.old
@@ -1,3 +1,3 @@
-2026/07/01-22:16:16.791 3008 Recovering log #1314
-2026/07/01-22:16:16.843 3008 Delete type=0 #1314
-2026/07/01-22:16:16.843 3008 Delete type=3 #1312
+2026/07/02-22:15:18.909 3304 Recovering log #1317
+2026/07/02-22:15:19.475 3304 Delete type=0 #1317
+2026/07/02-22:15:19.475 3304 Delete type=3 #1316
