+++ packs/equipamentos-cardigan/CURRENT
@@ -1,1 +1,1 @@
-MANIFEST-001329
+MANIFEST-001331
