--- packs/equipamentos-cardigan/LOG
@@ -1,3 +1,0 @@
-2026/07/02-22:15:22.146 25d0 Recovering log #1328
-2026/07/02-22:15:22.962 25d0 Delete type=0 #1328
-2026/07/02-22:15:22.962 25d0 Delete type=3 #1327
