+++ packs/equipamentos-cardigan/LOG.old
@@ -1,3 +1,3 @@
-2026/07/01-22:16:16.958 39a0 Recovering log #1325
-2026/07/01-22:16:17.020 39a0 Delete type=0 #1325
-2026/07/01-22:16:17.020 39a0 Delete type=3 #1323
+2026/07/02-22:15:22.146 25d0 Recovering log #1328
+2026/07/02-22:15:22.962 25d0 Delete type=0 #1328
+2026/07/02-22:15:22.962 25d0 Delete type=3 #1327
