+++ packs/racas-cardigan/CURRENT
@@ -1,1 +1,1 @@
-MANIFEST-002277
+MANIFEST-002279
