--- packs/racas-cardigan/LOG
@@ -1,3 +1,0 @@
-2026/07/02-22:15:21.275 2818 Recovering log #2276
-2026/07/02-22:15:22.092 2818 Delete type=0 #2276
-2026/07/02-22:15:22.092 2818 Delete type=3 #2275
