+++ packs/racas-cardigan/LOG.old
@@ -1,3 +1,3 @@
-2026/07/01-22:16:16.899 3798 Recovering log #2273
-2026/07/01-22:16:16.953 3798 Delete type=0 #2273
-2026/07/01-22:16:16.953 3798 Delete type=3 #2271
+2026/07/02-22:15:21.275 2818 Recovering log #2276
+2026/07/02-22:15:22.092 2818 Delete type=0 #2276
+2026/07/02-22:15:22.092 2818 Delete type=3 #2275
