+++ packs/skills-cardigan/CURRENT
@@ -1,1 +1,1 @@
-MANIFEST-001346
+MANIFEST-001348
