--- packs/skills-cardigan/LOG
@@ -1,3 +1,0 @@
-2026/07/02-22:15:19.525 2818 Recovering log #1345
-2026/07/02-22:15:20.935 2818 Delete type=0 #1345
-2026/07/02-22:15:20.935 2818 Delete type=3 #1344
