+++ packs/skills-cardigan/LOG.old
@@ -1,3 +1,3 @@
-2026/07/01-22:16:16.848 40cc Recovering log #1342
-2026/07/01-22:16:16.894 40cc Delete type=0 #1342
-2026/07/01-22:16:16.894 40cc Delete type=3 #1340
+2026/07/02-22:15:19.525 2818 Recovering log #1345
+2026/07/02-22:15:20.935 2818 Delete type=0 #1345
+2026/07/02-22:15:20.935 2818 Delete type=3 #1344
